Add dog file hole test and correct data type in migration

+++ db/schema.sql
@@ -8421,7 +8421,7 @@ CREATE TABLE maker.dog_file_hole (
     address_id bigint NOT NULL,
     header_id integer NOT NULL,
     what text,
-    data bigint
+    data numeric
 );
